fix: split provider files to satisfy backend-size gate
Extract OpenCode orchestration into inference_provider_opencode.go, fold provider error corpus into provider_errors.go, and relocate Codex stream and Claude failure parser tests into provider-owned packages so make backend-size passes without changing observable behavior.
