+++ pkg/workers/provider/provider_errors_test.go
@@ -2,7 +2,6 @@ package provider
 
 import (
 	"context"
-	"encoding/json"
 	"strings"
 	"testing"
 
@@ -60,121 +59,6 @@ func providerErrorCorpusLastErrorLine(t *testing.T, entry ProviderErrorCorpusEnt
 	return lastMatch
 }
 
-func TestParseClaudeProviderFailure_StructuredTypesAndStatusesAreCanonical(t *testing.T) {
-	testCases := []struct {
-		name        string
-		stderr      string
-		wantReason  interfaces.WorkFailureType
-		wantMessage string
-	}{
-		{
-			name:        "AuthenticationTypePreservesActionableMessage",
-			stderr:      `API Error: 500 {"type":"error","error":{"type":"authentication_error","message":"  Sign in again\nwith Claude Code.  "}}`,
-			wantReason:  interfaces.WorkFailureTypeAuthFailure,
-			wantMessage: "Sign in again with Claude Code.",
-		},
-		{
-			name:        "PermissionStatusFallbackPreservesActionableMessage",
-			stderr:      `API Error: 403 {"type":"error","error":{"message":"Ask an organization owner to grant access."}}`,
-			wantReason:  interfaces.WorkFailureTypeAuthFailure,
-			wantMessage: "Ask an organization owner to grant access.",
-		},
-		{
-			name:        "InvalidRequestPreservesCorrection",
-			stderr:      `API Error: 400 {"type":"error","error":{"type":"invalid_request_error","message":"Reduce the request size below 20 MB."}}`,
-			wantReason:  interfaces.WorkFailureTypePermanentBadRequest,
-			wantMessage: "Reduce the request size below 20 MB.",
-		},
-		{
-			name:        "RequestSizeStatusFallback",
-			stderr:      `API Error: 413 {"type":"error","error":{"message":"The request is too large; remove an attachment."}}`,
-			wantReason:  interfaces.WorkFailureTypePermanentBadRequest,
-			wantMessage: "The request is too large; remove an attachment.",
-		},
-		{
-			name:        "RateLimitUsesStableGuidance",
-			stderr:      `API Error: 429 {"type":"error","error":{"type":"rate_limit_error","message":"rate limit exceeded"}}`,
-			wantReason:  interfaces.WorkFailureTypeThrottled,
-			wantMessage: claudeThrottleFailureMessage,
-		},
-		{
-			name:        "OverloadTypeWinsOverConflictingStatus",
-			stderr:      `API Error: 401 {"type":"error","error":{"type":"overloaded_error","message":"overloaded"}}`,
-			wantReason:  interfaces.WorkFailureTypeThrottled,
-			wantMessage: claudeThrottleFailureMessage,
-		},
-		{
-			name:        "ServerStatusFallback",
-			stderr:      `API Error: 502 {"type":"error","error":{"message":"gateway included internal diagnostics"}}`,
-			wantReason:  interfaces.WorkFailureTypeInternalServerError,
-			wantMessage: claudeServerFailureMessage,
-		},
-	}
-
-	for _, tc := range testCases {
-		t.Run(tc.name, func(t *testing.T) {
-			got := ParseClaudeProviderFailure(CommandResult{ExitCode: 1, Stderr: []byte(tc.stderr)})
-			if got.Reason != tc.wantReason || got.Message != tc.wantMessage {
-				t.Fatalf("ParseClaudeProviderFailure() = %#v, want reason=%q message=%q", got, tc.wantReason, tc.wantMessage)
-			}
-		})
-	}
-}
-
-func TestParseClaudeProviderFailure_StructuredRecordPrecedesSurroundingText(t *testing.T) {
-	result := CommandResult{
-		ExitCode: 1,
-		Stderr: []byte(strings.Join([]string{
-			"cleanup mentioned 429 rate limit and forbidden credentials",
-			`API Error: 400 {"type":"error","error":{"type":"invalid_request_error","message":"Choose a model available to this project."}}`,
-			"post-command timeout while deleting a temporary directory",
-		}, "\n")),
-	}
-
-	got := ParseClaudeProviderFailure(result)
-	if got.Reason != interfaces.WorkFailureTypePermanentBadRequest || got.Message != "Choose a model available to this project." {
-		t.Fatalf("ParseClaudeProviderFailure() = %#v, want structured invalid-request result", got)
-	}
-}
-
-func TestParseClaudeProviderFailure_BoundsAndRejectsCredentialBearingMessages(t *testing.T) {
-	longCorrection := strings.Repeat("remove one attachment ", 100)
-	testCases := []struct {
-		name        string
-		message     string
-		wantMessage string
-	}{
-		{
-			name:        "BoundsNormalizedActionableDetail",
-			message:     longCorrection,
-			wantMessage: boundUTF8Bytes(strings.TrimSpace(longCorrection), claudeFailureMessageBytes),
-		},
-		{
-			name:        "CredentialSignalUsesProductFallback",
-			message:     "Replace Authorization: Bearer sk-ant-private",
-			wantMessage: claudeBadRequestFailureMessage,
-		},
-	}
-
-	for _, tc := range testCases {
-		t.Run(tc.name, func(t *testing.T) {
-			payload, err := json.Marshal(map[string]any{
-				"type": "error",
-				"error": map[string]string{
-					"type":    "invalid_request_error",
-					"message": tc.message,
-				},
-			})
-			if err != nil {
-				t.Fatalf("marshal fixture: %v", err)
-			}
-			got := ParseClaudeProviderFailure(CommandResult{ExitCode: 1, Stderr: []byte("API Error: 400 " + string(payload))})
-			if got.Message != tc.wantMessage || len(got.Message) > claudeFailureMessageBytes {
-				t.Fatalf("Message = %q (%d bytes), want %q bounded to %d bytes", got.Message, len(got.Message), tc.wantMessage, claudeFailureMessageBytes)
-			}
-		})
-	}
-}
 
 func TestNormalizeProviderExitFailure_ClaudeUsesOneParsedReasonAndMessageForPolicy(t *testing.T) {
 	entry := providerErrorCorpusEntryForTest(t, "claude_rate_limit_error")
@@ -192,105 +76,6 @@ func TestNormalizeProviderExitFailure_ClaudeUsesOneParsedReasonAndMessageForPoli
 	}
 }
 
-type claudeFallbackTestCase struct {
-	name        string
-	result      CommandResult
-	wantReason  interfaces.WorkFailureType
-	wantMessage string
-}
-
-func claudeFallbackTestCases() []claudeFallbackTestCase {
-	return []claudeFallbackTestCase{
-		{
-			name:        "TextConfigurationPreservesCorrection",
-			result:      CommandResult{ExitCode: 1, Stderr: []byte("Configuration error: set ANTHROPIC_BASE_URL to a valid URL.")},
-			wantReason:  interfaces.WorkFailureTypeMisconfigured,
-			wantMessage: "Configuration error: set ANTHROPIC_BASE_URL to a valid URL.",
-		},
-		{
-			name:        "TextAuthenticationPreservesCorrection",
-			result:      CommandResult{ExitCode: 1, Stderr: []byte("Not logged in. Run /login to continue.")},
-			wantReason:  interfaces.WorkFailureTypeAuthFailure,
-			wantMessage: "Not logged in. Run /login to continue.",
-		},
-		{
-			name:        "MalformedEnvelopeFallsThroughToTextSignal",
-			result:      CommandResult{ExitCode: 1, Stderr: []byte(`API Error: 429 {"type":"error","error":{"type":"rate_limit_error"`)},
-			wantReason:  interfaces.WorkFailureTypeThrottled,
-			wantMessage: claudeThrottleFailureMessage,
-		},
-		{
-			name: "LastRecognizedTextRecordExcludesCleanup",
-			result: CommandResult{ExitCode: 1, Stderr: []byte(strings.Join([]string{
-				"rate limit exceeded",
-				"Invalid request: select a supported Claude model.",
-				"cleanup finished after timeout directory removal",
-			}, "\n"))},
-			wantReason:  interfaces.WorkFailureTypePermanentBadRequest,
-			wantMessage: "Invalid request: select a supported Claude model.",
-		},
-		{
-			name:        "TimeoutExitUsesStableGuidance",
-			result:      CommandResult{ExitCode: 124},
-			wantReason:  interfaces.WorkFailureTypeTimeout,
-			wantMessage: claudeTimeoutFailureMessage,
-		},
-		{
-			name:        "SingleSafeUnknownUsesBoundedExcerpt",
-			result:      CommandResult{ExitCode: 9, Stderr: []byte("some brand new claude failure")},
-			wantReason:  interfaces.WorkFailureTypeUnknown,
-			wantMessage: "Claude failed: some brand new claude failure",
-		},
-		{
-			name:        "NoisyUnknownUsesExitCodeFallback",
-			result:      CommandResult{ExitCode: 7, Stderr: []byte("User: private prompt\ncleanup complete")},
-			wantReason:  interfaces.WorkFailureTypeUnknown,
-			wantMessage: "claude exited with code 7",
-		},
-		{
-			name:        "CredentialUnknownUsesExitCodeFallback",
-			result:      CommandResult{ExitCode: 2, Stderr: []byte("Authorization: Bearer sk-ant-private")},
-			wantReason:  interfaces.WorkFailureTypeUnknown,
-			wantMessage: "claude exited with code 2",
-		},
-	}
-}
-
-func TestParseClaudeProviderFailure_TextMalformedNoisyAndUnknownOutcomes(t *testing.T) {
-	for _, tc := range claudeFallbackTestCases() {
-		t.Run(tc.name, func(t *testing.T) {
-			got := ParseClaudeProviderFailure(tc.result)
-			if got.Reason != tc.wantReason || got.Message != tc.wantMessage {
-				t.Fatalf("ParseClaudeProviderFailure() = %#v, want reason=%q message=%q", got, tc.wantReason, tc.wantMessage)
-			}
-			if len(got.Message) > claudeFailureMessageBytes {
-				t.Fatalf("Message length = %d, want at most %d bytes", len(got.Message), claudeFailureMessageBytes)
-			}
-		})
-	}
-}
-
-func TestProviderErrorCorpus_ClaudeFixturesMatchCanonicalParserAndPolicy(t *testing.T) {
-	corpus := loadProviderErrorCorpusForTest(t)
-	for _, entry := range corpus.allEntries {
-		if entry.Provider != interfaces.ModelProviderClaude {
-			continue
-		}
-		t.Run(entry.Name, func(t *testing.T) {
-			parsed := ParseClaudeProviderFailure(entry.CommandResult())
-			if parsed.Reason != entry.ExpectedType || parsed.Message != entry.ExpectedMessage {
-				t.Fatalf("parsed = %#v, want reason=%q message=%q", parsed, entry.ExpectedType, entry.ExpectedMessage)
-			}
-			providerErr := NewProviderErrorFromResult(parsed, nil)
-			decision := WorkFailureDecisionFromProviderError(providerErr)
-			if providerErr.Family != entry.ExpectedFamily ||
-				decision.Retryable != entry.Retryable ||
-				decision.TriggersThrottlePause != entry.TriggersThrottlePause {
-				t.Fatalf("policy = family %q decision %#v, want family %q retryable=%t throttle=%t", providerErr.Family, decision, entry.ExpectedFamily, entry.Retryable, entry.TriggersThrottlePause)
-			}
-		})
-	}
-}
 
 func TestParseCodexProviderFailure_GPT56SolReturnsActionableNestedMessage(t *testing.T) {
 	entry := providerErrorCorpusEntryForTest(t, "codex_gpt_5_6_sol_requires_newer_cli")
